+++ config/lerk.rb
@@ -80,6 +80,7 @@ EOF
       HTTP_BASIC_PASSWORD = ENV['GORGE_HTTP_BASIC_PASSWORD']
       CONNECT_TIMEOUT     = ENV.fetch('GORGE_CONNECT_TIMEOUT', 1).to_i
       TIMEOUT             = ENV.fetch('GORGE_TIMEOUT', 2).to_i
+      STATISTICS_CLASS    = ENV.fetch('GORGE_STATISTICS_CLASS', 'n_30')
     end
 
     module Statistics