+++ CHANGELOG.md
@@ -8,6 +8,8 @@ glance - what to expact from upgrading to a new version.
 
 ### Added
 
+- Support for Gorge's cached player statistics.
+
 ### Changed
 
 ### Fixed